update design and code to use the Nano board and more compact pin layout change ServoExt data types to fix integer operations update LCD pins to the layout that works with the I2C board

+++ SelfDrivingCar/ServoExt.h
@@ -20,40 +20,39 @@ class ServoExt : public Servo
      * @param maxAngle maximum angle that can be asked from this servo
      * @param travelTime time in milliseconds to reach the maximum angle, starting from the middle point
      */
-    ServoExt(const int8_t maxAngle, const int8_t travelTime);
-    
+    ServoExt(const unsigned short int maxAngle, const unsigned short int travelTime);
+
     /**
      * Set a percentage of travel relative to the center, eg [-100, 100]
      * @param percentage percentage of travel to set
      * @param waitToStabilize whether or not to wait for the expected time of the servo to reach the destination
      */
-    void setPercentage(const int8_t percentage, const bool waitToStabilize = false);
-    
+    void setPercentage(const unsigned short int percentage, const bool waitToStabilize = false);
+
     /**
      * Set the desired angle relative to the center, eg [-90, 90]. It will automatically apply the constructor-set boundary
      * @param angle desired angle
      * @param waitToStabilize whether or not to wait for the expected time of the servo to reach the destination
      */
-    void setAngle(int8_t angle, const bool waitToStabilize = false);
-    
+    void setAngle(unsigned short int angle, const bool waitToStabilize = false);
+
     /**
      * Wait long enough for the servo to reach the point of the last operation (called with waitToStabilize = false)
      */
     void waitForAngle();
-    
+
     /**
      * Get the angle last set by one of the above operations
      */
-    int8_t getAngle();
+    unsigned short int getAngle();
 
   private:
-    int8_t maxAngle;
-    int8_t travelTime;
-    int8_t oldAngle;
-    int8_t prevAngle;
+    unsigned short int maxAngle;
+    unsigned short int travelTime;
+    unsigned short int oldAngle;
+    unsigned short int prevAngle;
 
     unsigned long toStabilize;
 };
 
 #endif
-
